move global action creator to it's own folder
